+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -4,7 +4,7 @@ public class StringUtils {
     public static boolean isPositiveNumber(String str) {
          //here magic will happen
         try{
-            return Integer.parseInt(str) >0;}
+            return Float.parseFloat(str) >0;}
     catch (Exception e){
         return false;}
     }}
